requirements.txt
amqp==5.0.1
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.2
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cloudinary==1.23.0
config==0.5.0.post0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.7.0
django-heroku==0.3.1
django-nested-admin==3.3.3
django-templated-mail==1.1.1
djangorestframework==3.12.1
django-phonenumber-field ==5.0.0
djoser==2.0.5
filelock==3.0.12
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
phonenumbers==8.12.11
Pillow==6.2.0
pipenv==2020.11.15
promise==2.3
prompt-toolkit==3.0.7
psycopg2==2.8.6
pycodestyle==2.6.0
python-dotenv==0.14.0
python-monkey-business==1.0.0
python-decouple==3.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
Rx==1.6.1
sentry-sdk==0.19.4
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.4.1
toml==0.10.2
twilio==6.46.0
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
